Risenholm: add experimental SetPCLikeStatus
The idea here is that we can set the PC like status on the player list
without changing their actual hostility. So when we add in invaders, we
can make the invading PC hostile to the PCs they're spicing, without
revealing on the player list that they're the one who's doing it.

We'll also need to use the feedback plugin to prevent the "dislikes you"
message.